Robot init

+++ 18x18 VEX-U 2025/include/robot-config.hpp
@@ -6,7 +6,7 @@
 
 
 struct MotorConfig {
-    int PORT_NUMBER;
+    int8_t PORT_NUMBER;
     pros::v5::MotorGears GEARSET;
     pros::v5::MotorUnits MOTOR_UNITS;
     bool REVERSED;
